pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "telegraph"
version = "0.1.0"
description = "CLI tool to post content to telegra.ph"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "mistune>=3.0",
    "beautifulsoup4>=4.12",
    "rich>=13.0",
    "PyYAML>=6.0",
]

[project.scripts]
telegraph = "telegraph.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/telegraph"]
